Move timespan conversion into TimeSpanConverter.
